# docs/swagger.yaml
basePath: /v1
definitions:
  models.Mountain:
    properties:
      created_at:
        type: string
      elevation_m:
        type: integer
      id:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      province:
        type: string
      routes:
        items:
          $ref: '#/definitions/models.Route'
        type: array
      updated_at:
        type: string
    type: object
  models.Post:
    properties:
      altitude_m:
        type: integer
      created_at:
        type: string
      distance_from_start_m:
        type: integer
      distance_to_summit_m:
        type: integer
      id:
        type: string
      name:
        type: string
      order_index:
        type: integer
      route_id:
        type: string
      segment_to_next_m:
        type: integer
      updated_at:
        type: string
    type: object
  models.Route:
    properties:
      created_at:
        type: string
      difficulty:
        type: string
      elevation_gain_m:
        type: integer
      id:
        type: string
      is_official:
        type: boolean
      length_m:
        type: integer
      mountain_id:
        type: string
      name:
        type: string
      posts:
        items:
          $ref: '#/definitions/models.Post'
        type: array
      start_point_name:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for Mountain Hiking Data in Indonesia
  title: Mountain API
  version: "1.0"
paths:
  /mountains:
    get:
      description: Get a list of mountains with optional filtering
      parameters:
      - description: Search query
        in: query
        name: query
        type: string
      - description: Filter by province
        in: query
        name: province
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Mountain'
            type: array
      summary: Get all mountains
      tags:
      - mountains
    post:
      consumes:
      - application/json
      description: Create a new mountain record
      parameters:
      - description: Mountain Data
        in: body
        name: mountain
        required: true
        schema:
          $ref: '#/definitions/models.Mountain'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Mountain'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new mountain
      tags:
      - mountains
  /mountains/{id}:
    delete:
      description: Delete a mountain and its routes/posts
      parameters:
      - description: Mountain ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete a mountain
      tags:
      - mountains
    get:
      description: Get details of a specific mountain
      parameters:
      - description: Mountain ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mountain'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a mountain by ID
      tags:
      - mountains
    patch:
      consumes:
      - application/json
      description: Update specific fields of a mountain
      parameters:
      - description: Mountain ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: mountain
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mountain'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a mountain partially
      tags:
      - mountains
    put:
      consumes:
      - application/json
      description: Update all fields of a mountain
      parameters:
      - description: Mountain ID
        in: path
        name: id
        required: true
        type: string
      - description: Mountain Data
        in: body
        name: mountain
        required: true
        schema:
          $ref: '#/definitions/models.Mountain'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Mountain'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update a mountain fully
      tags:
      - mountains
  /mountains/{id}/routes:
    get:
      description: Get all routes associated with a mountain
      parameters:
      - description: Mountain ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Route'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get routes for a mountain
      tags:
      - mountains
  /posts:
    post:
      consumes:
      - application/json
      parameters:
      - description: Post Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Create a new post
      tags:
      - posts
  /posts/{id}:
    delete:
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete a post
      tags:
      - posts
    patch:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: post
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Update a post partially
      tags:
      - posts
    put:
      consumes:
      - application/json
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post Data
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
      security:
      - ApiKeyAuth: []
      summary: Update a post fully
      tags:
      - posts
  /posts/bulk:
    post:
      consumes:
      - multipart/form-data
      description: Upload multiple posts given a route_id via CSV
      parameters:
      - description: Route ID
        in: formData
        name: route_id
        required: true
        type: string
      - description: CSV File (name,order_index,altitude_m,distance_to_summit_m,...)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Bulk upload posts from CSV
      tags:
      - posts
  /routes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Route Data
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/models.Route'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Route'
      security:
      - ApiKeyAuth: []
      summary: Create a new route
      tags:
      - routes
  /routes/{id}:
    delete:
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - ApiKeyAuth: []
      summary: Delete a route
      tags:
      - routes
    get:
      description: Get details of a specific route
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Route'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a route by ID
      tags:
      - routes
    patch:
      consumes:
      - application/json
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: route
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Route'
      security:
      - ApiKeyAuth: []
      summary: Update a route partially
      tags:
      - routes
    put:
      consumes:
      - application/json
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      - description: Route Data
        in: body
        name: route
        required: true
        schema:
          $ref: '#/definitions/models.Route'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Route'
      security:
      - ApiKeyAuth: []
      summary: Update a route fully
      tags:
      - routes
  /routes/{id}/posts:
    get:
      description: Get all posts associated with a route, ordered by index
      parameters:
      - description: Route ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Post'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get posts for a route
      tags:
      - routes
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"